+++ src/app/orders/interfaces/ioder.ts
@@ -1,6 +1,8 @@
 import { IOrderItem } from './iorderitem';
+import { IUser } from 'src/app/core/models/iuser';
 
 export interface IOrder {
     orderItems: IOrderItem[];
+    user: IUser;
     id: number;
 }